requirements.txt
aiohttp==3.8.1
asyncer==0.0.2
click==8.1.3
fastapi==0.87.0
filetype==1.2.0
pooch==1.6.0
imagehash==4.3.1
numpy
onnxruntime==1.13.1
pillow
pymatting==1.1.8
python-multipart==0.0.5
uvicorn==0.20.0
watchdog==2.1.9
